prediction confidently incorrect

+++ esp_tflite/components/audio_process/include/micro_features_generator.h
@@ -19,7 +19,7 @@ limitations under the License.
 #include "tensorflow/lite/c/common.h"
 #include "micro_model_settings.h"
 
-using Features = int8_t[kFeatureCount][kFeatureSize];
+using Features = float[kFeatureCount][kFeatureSize];
 
 // Sets up any resources needed for the feature generation pipeline.
 TfLiteStatus InitializeMicroFeatures();
@@ -30,8 +30,8 @@ TfLiteStatus InitializeMicroFeatures();
 //                                    int output_size, int8_t* output,
 //                                    size_t* num_samples_read);
 
-TfLiteStatus GenerateFeatures(const int16_t* audio_data,
+TfLiteStatus GenerateFeatures(const int16_t *audio_data,
                               const size_t audio_data_size,
-                              Features* features_output);
+                              Features *features_output);
 
-#endif  // TENSORFLOW_LITE_MICRO_EXAMPLES_MICRO_SPEECH_MICRO_FEATURES_MICRO_FEATURES_GENERATOR_H_
+#endif // TENSORFLOW_LITE_MICRO_EXAMPLES_MICRO_SPEECH_MICRO_FEATURES_MICRO_FEATURES_GENERATOR_H_
